feat(session): extract managed-session naming to trusty-common; trusty-agents adopts it (DOC-33 Phase 1, SPEC-ONESM-01)
Phase 1 of the one-session-manager consolidation (#1987). trusty-agents' TmManager
named sessions <project>-<harness>-<serial> (no managed prefix), so they were
invisible to trusty-mpm's reconcile/prune/adopt (which filter on the managed
prefix) — orphaned, violating the no-fire-and-forget standard.

- Move the canonical session-naming module (name_from_uuid/name_from_dir/
  build_session_name/is_managed_session_name + the tm-/tmpm-/trusty-mpm- prefix
  recognition, 40 unit tests) from trusty-mpm::core::names to
  trusty_common::session_naming, behind a `session-naming` feature.
- trusty-mpm::core::names becomes a thin `pub use` facade — every existing caller
  and test resolves unchanged.
- trusty-agents' next_session_name now delegates to the shared build_session_name,
  emitting a managed-prefixed tmux name (e.g. tm-trusty-agents-repl-01) that
  is_managed_session_name recognizes. This fixes the orphaned-session symptom.

trusty-common/trusty-mpm/trusty-agents test+clippy+fmt+line-cap green.
